Visualising the Central Limit Theorem.
If you ever wondered how does the CLT really work, but didn't find the
mathematical proof enough, then you can try this Python script and see
for yourself how does the distribution of the sample of means change in
response to the change in the size of each sample. You can also vary the
number of the samples.